[MNG-5760] Extract public interface
